implement full redux support (i.e. management of external state)
