+++ src/navigationStore.js
@@ -1,7 +1,20 @@
+import React from 'react';
 import { observable, action, autorunAsync } from 'mobx';
-import { NavigationActions } from 'react-navigation';
 import * as ActionConst from './ActionConst';
-import { OnEnter, OnExit } from './Util';
+import { OnEnter, OnExit, assert } from './Util';
+import { View } from 'react-native';
+import { TabNavigator, DrawerNavigator, StackNavigator, NavigationActions } from 'react-navigation';
+import { LeftButton, RightButton, renderBackButton } from './NavBar';
+import LightboxNavigator from './LightboxNavigator';
+import OverlayNavigator from './OverlayNavigator';
+import _drawerImage from '../images/menu_burger.png';
+import Scene from './Scene';
+import PropTypes from 'prop-types';
+import { getActiveState } from './State';
+import { reducer } from './Reducer';
+
+let RightNavBarButton;
+let LeftNavBarButton;
 
 export const actionMap = {
   [ActionConst.JUMP]: 'jump',
@@ -17,14 +30,175 @@ export const actionMap = {
   [ActionConst.POP_AND_PUSH]: 'popAndPush',
 };
 
-export const supportedActions = {
-  [ActionConst.PUSH]: NavigationActions.NAVIGATE,
-  [ActionConst.JUMP]: NavigationActions.NAVIGATE,
-  [ActionConst.BACK]: NavigationActions.BACK,
-  [ActionConst.REFRESH]: NavigationActions.BACK,
-  [ActionConst.RESET]: NavigationActions.RESET,
-  [ActionConst.REPLACE]: NavigationActions.RESET,
-};
+const reservedKeys = [
+  'children',
+  'create',
+  'execute',
+  'popTo',
+  'navigate',
+  'replace',
+  'refresh',
+  'dispatch',
+  'push',
+  'setParams',
+  'run',
+  'onEnter',
+  'onRight',
+  'onLeft',
+  'left',
+  'back',
+  'right',
+  'rightButton',
+  'leftButton',
+  'on',
+  'onExit',
+  'pop',
+  'renderLeftButton',
+  'renderRightButton',
+  'renderTitle',
+  'navBar',
+  'title',
+  'drawerOpen',
+  'drawerClose',
+];
+
+const dontInheritKeys = [
+  'component',
+  'wrap',
+  'modal',
+  'overlay',
+  'drawer',
+  'tabs',
+  'navigator',
+  'children',
+  'key',
+  'ref',
+  'style',
+  'title',
+  'navTransparent',
+  'type',
+  'hideNavBar',
+  'hideTabBar',
+];
+
+function getValue(value, params) {
+  return value instanceof Function ? value(params) : value;
+}
+
+function getProperties(component = {}) {
+  const res = {};
+  for (const key of reservedKeys) {
+    if (component[key]) {
+      res[key] = component[key];
+    }
+  }
+  delete res.children;
+  return res;
+}
+function createTabBarOptions({ tabBarStyle, activeTintColor, inactiveTintColor, activeBackgroundColor, inactiveBackgroundColor, showLabel, labelStyle, tabStyle, ...props }) {
+  return { ...props, style: tabBarStyle, activeTintColor, inactiveTintColor, activeBackgroundColor, inactiveBackgroundColor, showLabel, labelStyle, tabStyle };
+}
+function createNavigationOptions(params) {
+  const { title, backButtonImage, navTransparent, hideNavBar, hideTabBar, backTitle, right, rightButton, left, leftButton,
+    navigationBarStyle, headerStyle, navBarButtonColor, tabBarLabel, tabBarIcon, icon, getTitle, renderTitle, panHandlers,
+    navigationBarTitleImage, navigationBarTitleImageStyle, component, rightTitle, leftTitle, leftButtonTextStyle, rightButtonTextStyle,
+    backButtonTextStyle, headerTitleStyle, titleStyle, navBar, onRight, onLeft, rightButtonImage, leftButtonImage, init, back, ...props } = params;
+  const NavBar = navBar;
+  if (component && component.navigationOptions) {
+    return component.navigationOptions;
+  }
+  return ({ navigation, screenProps }) => {
+    const navigationParams = navigation.state.params || {};
+    const state = { navigation, ...params, ...navigationParams, ...screenProps };
+    const res = {
+      ...props,
+      headerTintColor: navBarButtonColor || props.tintColor || navigationParams.tintColor || navigationParams.headerTintColor,
+      headerTitleStyle: headerTitleStyle || titleStyle,
+      title: getValue((navigationParams.title) || title || getTitle, state),
+      headerBackTitle: getValue((navigationParams.backTitle) || backTitle, state),
+      headerRight: getValue((navigationParams.right) || right || rightButton || params.renderRightButton, state),
+      headerLeft: getValue((navigationParams.left) || left || leftButton || params.renderLeftButton, state),
+      headerTitle: getValue((navigationParams.renderTitle) || renderTitle || params.renderTitle, state),
+      headerStyle: getValue((navigationParams.headerStyle || headerStyle || navigationBarStyle), state),
+      headerBackImage: navigationParams.backButtonImage || backButtonImage,
+    };
+    if (NavBar) {
+      res.header = (data) => <NavBar navigation={navigation} {...state} {...data} />;
+    }
+
+    if (panHandlers === null) {
+      res.gesturesEnabled = false;
+    }
+
+    if (navigationBarTitleImage) {
+      res.headerTitle = <Image source={navigationBarTitleImage} style={navigationBarTitleImageStyle} />;
+    }
+
+    if (tabBarLabel) {
+      res.tabBarLabel = tabBarLabel;
+    }
+
+    if (tabBarIcon || icon) {
+      res.tabBarIcon = tabBarIcon || icon;
+    }
+    const componentData = {};
+    // copy all component static functions
+    if (component) {
+      for (const key of ['onRight', 'onLeft', 'rightButton', 'leftButton', 'leftTitle', 'rightTitle', 'rightButtonImage',
+        'leftButtonImage', 'rightButtonTextStyle', 'leftButtonTextStyle', 'rightButtonIconStyle', 'leftButtonIconStyle',
+        'leftButtonTintColor', 'rightButtonTintColor']) {
+        if (component[key]) {
+          componentData[key] = component[key];
+        }
+      }
+    }
+
+    if (rightButtonImage || rightTitle || params.renderRightButton || onRight || navigationParams.onRight
+      || navigationParams.rightTitle || navigationParams.rightButtonImage || rightButtonTextStyle) {
+      res.headerRight = getValue(navigationParams.right || navigationParams.rightButton || params.renderRightButton,
+        { ...navigationParams, ...screenProps }) || <RightNavBarButton {...params} {...navigationParams} {...componentData} />;
+    }
+
+    if (leftButtonImage || backButtonImage || backTitle || leftTitle || params.renderLeftButton || leftButtonTextStyle
+      || backButtonTextStyle || onLeft || navigationParams.leftTitle || navigationParams.onLeft || navigationParams.leftButtonImage
+      || navigationParams.backButtonImage || navigationParams.backTitle) {
+      res.headerLeft = getValue(navigationParams.left || navigationParams.leftButton || params.renderLeftButton, { ...params, ...navigationParams, ...screenProps })
+        || (onLeft && (leftTitle || navigationParams.leftTitle || leftButtonImage) && <LeftNavBarButton {...params} {...navigationParams} {...componentData} />)
+        || (init ? null : renderBackButton({ ...params, ...navigationParams, ...screenProps }));
+    }
+
+    if (back) {
+      res.headerLeft = renderBackButton({ ...params, ...navigationParams });
+    }
+
+    if (hideTabBar) {
+      res.tabBarVisible = false;
+    }
+    if (hideNavBar) {
+      res.header = null;
+    }
+
+    if (navTransparent) {
+      res.headerStyle = { position: 'absolute', backgroundColor: 'transparent', zIndex: 100, top: 0, left: 0, right: 0 };
+    }
+    return res;
+  };
+}
+
+function createWrapper(Component, wrapBy) {
+  if (!Component) {
+    return null;
+  }
+  const wrapper = wrapBy || (props => props);
+  function wrapped({ navigation, ...props }) {
+    return <Component {...props} navigation={navigation} {...navigation.state.params} name={navigation.state.routeName} />;
+  }
+  wrapped.propTypes = {
+    navigation: PropTypes.object,
+  };
+  return wrapper(wrapped);
+}
+
 function filterParam(data) {
   if (data.toString() !== '[object Object]') {
     return { data };
@@ -48,16 +222,10 @@ function uniteParams(routeName, params) {
   return res;
 }
 
-const createAction = (type: string) => (payload: Object = {}) => ({
-  type,
-  ...payload,
-});
-
-
 class NavigationStore {
-  _router = null;
   states = {};
   reducer = null;
+  router;
   _state;
   @observable currentScene = '';
   @observable prevScene = '';
@@ -69,14 +237,6 @@ class NavigationStore {
     const scene = this.currentScene;// eslint-disable-line no-unused-vars
     const params = this.currentParams;// eslint-disable-line no-unused-vars
     return this._state;
-  }
-
-  set router(router) {
-    this._router = router;
-    this.dispatch(NavigationActions.init());
-  }
-  get router() {
-    return this._router;
   }
 
   constructor() {
@@ -108,7 +268,7 @@ class NavigationStore {
           // call onEnter handler
           if (handler) {
             try {
-              const params = this.currentState().params;
+              const params = getActiveState(this._state).params;
               const res = await handler(params);
               if (res) {
                 success(res);
@@ -126,43 +286,164 @@ class NavigationStore {
     });
   }
 
-  nextState = (state, cmd) => (this.reducer ? this.reducer(state, cmd) : this._router.getStateForAction(cmd, state));
-
-  dispatch = (cmd, type, params) => {
-    this.setState(this.nextState(this.state, cmd), type, params);
+  create = (scene: Scene, params = {}, wrapBy = props => props) => {
+    assert(!Array.isArray(scene), 'Router should contain only one scene, please wrap your scenes with root Scene ');
+    RightNavBarButton = wrapBy(RightButton);
+    LeftNavBarButton = wrapBy(LeftButton);
+    const AppNavigator = this.processScene(scene, params, [], wrapBy);
+    this.router = AppNavigator.router;
+    this.dispatch(NavigationActions.init());
+    return AppNavigator;
   };
 
-  @action setState = (newState, type, params) => {
+  processScene = (scene: Scene, inheritProps = {}, clones = [], wrapBy) => {
+    assert(scene.props, 'props should be defined');
+    if (!scene.props.children) {
+      return null;
+    }
+    const res = {};
+    const order = [];
+    const { tabs, modal, overlay, lightbox, navigator, contentComponent, lazy, drawer, ...parentProps } = scene.props;
+
+    const commonProps = { ...inheritProps, ...parentProps };
+    delete commonProps.children;
+    delete commonProps.component;
+    // add inherit props
+    for (const pkey of Object.keys(commonProps)) {
+      if (dontInheritKeys.includes(pkey) && !parentProps[pkey]) {
+        delete commonProps[pkey];
+      }
+    }
+
+    if (drawer && commonProps.drawerPosition !== 'right' && !commonProps.left && !commonProps.leftButtonImage
+      && !commonProps.leftTitle && !commonProps.back) {
+      commonProps.leftButtonImage = commonProps.drawerImage || _drawerImage;
+      commonProps.onLeft = this.drawerOpen;
+    }
+
+    if (drawer && commonProps.drawerPosition === 'right' && !commonProps.right && !commonProps.rightButtonImage
+      && !commonProps.rightTitle) {
+      commonProps.rightButtonImage = commonProps.drawerImage || _drawerImage;
+      commonProps.onRight = this.drawerOpen;
+    }
+
+    const children = !Array.isArray(parentProps.children) ? [parentProps.children] : [...parentProps.children];
+    // add clone scenes
+    if (!drawer && !tabs) {
+      children.push(...clones);
+    }
+    // add all clones
+    for (const child of children) {
+      if (child.props.clone) {
+        if (clones.indexOf(child) === -1) {
+          clones.push(child);
+        }
+      }
+    }
+    let initialRouteName;
+    let initialRouteParams;
+    for (const child of children) {
+      assert(child.key, `key should be defined for ${child}`);
+      const key = child.key;
+      const init = key === children[0].key;
+      assert(reservedKeys.indexOf(key) === -1, `Scene name cannot be reserved word: ${child.key}`);
+      const { component, type = tabs || drawer ? 'jump' : 'push', onEnter, onExit, on, failure, success, wrap, ...props } = child.props;
+      if (!this.states[key]) {
+        this.states[key] = {};
+      }
+      for (const transition of Object.keys(props)) {
+        if (reservedKeys.indexOf(transition) === -1 && (props[transition] instanceof Function)) {
+          this.states[key][transition] = props[transition];
+        }
+      }
+      delete props.children;
+      if (success) {
+        this.states[key].success = success instanceof Function ?
+          success : args => { console.log(`Transition to state=${success}`); this[success](args); };
+      }
+      if (failure) {
+        this.states[key].failure = failure instanceof Function ?
+          failure : args => { console.log(`Transition to state=${failure}`); this[failure](args); };
+      }
+      // console.log(`KEY ${key} DRAWER ${drawer} TABS ${tabs} WRAP ${wrap}`, JSON.stringify(commonProps));
+      const screen = {
+        screen: createWrapper(component, wrapBy) || this.processScene(child, commonProps, clones) || (lightbox && View),
+        navigationOptions: createNavigationOptions({ ...commonProps, ...getProperties(component), ...child.props, init, component }),
+      };
+
+      // wrap component inside own navbar for tabs/drawer parent controllers
+      const wrapNavBar = drawer || tabs || wrap;
+      if (component && wrapNavBar) {
+        res[key] = { screen: this.processScene({ key, props: { children: { key: `_${key}`, props: { ...child.props, wrap: false } } } }, commonProps, clones, wrapBy) };
+      } else {
+        res[key] = screen;
+      }
+
+      // a bit of magic, create all 'actions'-shortcuts inside navigationStore
+      props.init = true;
+      if (!this[key]) {
+        this[key] = new Function('actions', 'props', 'type', // eslint-disable-line no-new-func
+          `return function ${key}(params){ actions.execute(type, '${key}', props, params)}`)(this, { ...commonProps, ...props }, type);
+      }
+
+      if ((onEnter || on || (component && component.onEnter)) && !this[key + OnEnter]) {
+        this[key + OnEnter] = onEnter || on || component.onEnter;
+      }
+
+      if ((onExit || (component && component.onExit)) && !this[key + OnExit]) {
+        this[key + OnExit] = onExit || component.onExit;
+      }
+
+      order.push(key);
+      if (child.props.initial || !initialRouteName) {
+        initialRouteName = key;
+        initialRouteParams = { ...commonProps, ...props };
+      }
+    }
+    const mode = modal ? 'modal' : 'card';
+    if (navigator) {
+      return navigator(res, { lazy, initialRouteName, initialRouteParams, contentComponent, order, ...commonProps, navigationOptions: createNavigationOptions(commonProps) });
+    }
+    if (lightbox) {
+      return LightboxNavigator(res, { mode, initialRouteParams, initialRouteName, ...commonProps, navigationOptions: createNavigationOptions(commonProps) });
+    } else if (tabs) {
+      return TabNavigator(res, { lazy, initialRouteName, initialRouteParams, order, ...commonProps,
+        tabBarOptions: createTabBarOptions(commonProps), navigationOptions: createNavigationOptions(commonProps) });
+    } else if (drawer) {
+      return DrawerNavigator(res, { initialRouteName, contentComponent, order, ...commonProps });
+    } else if (overlay) {
+      return OverlayNavigator(res, { mode, initialRouteParams, order, contentComponent, initialRouteName, ...commonProps,
+        navigationOptions: createNavigationOptions(commonProps) });
+    }
+    return StackNavigator(res, { mode, initialRouteParams, initialRouteName, ...commonProps, navigationOptions: createNavigationOptions(commonProps) });
+  };
+
+  nextState = (state, cmd) => (this.reducer ? this.reducer(state, cmd) : reducer(state, cmd));
+
+  dispatch = (cmd) => {
+    this.setState(this.nextState(this.state, cmd));
+  };
+
+  @action setState = (newState) => {
     // don't allow null state
     if (!newState) {
       return;
     }
-    const state = this.currentState(newState);
-    if (type === ActionConst.JUMP && state.routeName === this.currentScene) {
-      return;
-    }
+    const state = getActiveState(newState);
+    const prevScene = this.prevScene;
+    const currentScene = this.currentScene;
     this._state = newState;
-    // run 'blur' event
-    if (this.reducer) {
-      const overridenState = this.reducer(newState, { type: ActionConst.BLUR, routeName: this.prevScene });
-      if (overridenState) {
-        this._state = overridenState;
-      }
+    if (prevScene !== this.prevScene) {
+      this.dispatch({ type: ActionConst.BLUR, routeName: this.prevScene });
     }
     this.prevScene = this.currentScene;
     this._onExitHandlerExecuted = false;
     this._onEnterHandlerExecuted = false;
     this.currentScene = state.routeName;
     this.currentParams = state.params;
-    if (type === ActionConst.JUMP && params) {
-      this.refresh(params);
-    }
-    // run 'focus' event
-    if (this.reducer) {
-      const overridenState = this.reducer(newState, { type: ActionConst.FOCUS, routeName: this.currentScene, params: state.params });
-      if (overridenState) {
-        this._state = overridenState;
-      }
+
+    if (currentScene !== this.currentScene) {
+      this.dispatch({ type: ActionConst.FOCUS, routeName: this.currentScene, params: this.currentParams });
     }
   };
 
@@ -173,45 +454,14 @@ class NavigationStore {
     this[type](routeName, res);
   };
 
-  run = (type = ActionConst.PUSH, routeName, actions = {}, ...params) => {
-    const res = uniteParams(routeName, params);
-    if (supportedActions[type]) {
-      this.dispatch(createAction(supportedActions[type])({ routeName, ...actions, params: res }), type, res);
-    } else {
-      if (type === ActionConst.POP_TO) {
-        let nextScene = '';
-        let newState = this._state;
-        let currentState = this._state;
-        const currentScene = this.currentScene;
-        while (nextScene !== currentScene && newState && nextScene !== routeName) {
-          newState = this.nextState(currentState, NavigationActions.back());
-          if (newState) {
-            nextScene = this.currentState(newState).routeName;
-            if (nextScene !== routeName) {
-              currentState = newState;
-            }
-          }
-        }
-        if (nextScene === routeName) {
-          this.setState(newState);
-        }
-      } else if (type === ActionConst.POP_AND_PUSH) {
-        this.pop();
-        this.push(routeName, ...params);
-      }
-      // pass to reducer to notify
-      if (this.reducer) {
-        this.setState(this.reducer(this.state, { type, routeName, params: res }));
-      }
-    }
+  push = (routeName, data) => {
+    const params = filterParam(data);
+    this.dispatch({ type: ActionConst.PUSH, routeName, params });
   };
 
-  push = (routeName, ...params) => {
-    this.run(ActionConst.PUSH, routeName, null, ...params);
-  };
-
-  jump = (routeName, ...params) => {
-    this.run(ActionConst.JUMP, routeName, null, ...params);
+  jump = (routeName, data) => {
+    const params = filterParam(data);
+    this.dispatch({ type: ActionConst.JUMP, routeName, params });
   };
 
   drawerOpen = () => {
@@ -222,19 +472,9 @@ class NavigationStore {
     this.dispatch(NavigationActions.navigate({ routeName: 'DrawerClose' }));
   };
 
-  currentState = (param) => {
-    let state = param;
-    if (!state) {
-      state = this._state;
-    }
-    if (!state.routes) {
-      return state;
-    }
-    return this.currentState(state.routes[state.index]);
-  };
-
-  refresh = (params) => {
-    const key = this.currentState(this.state).key;
+  refresh = (data) => {
+    const key = getActiveState(this._state).key;
+    const params = filterParam(data);
     this.dispatch(NavigationActions.setParams({ key, params }));
   };
 
@@ -246,28 +486,30 @@ class NavigationStore {
     }
   };
 
-  popTo = (routeName, ...params) => {
-    this.run(ActionConst.POP_TO, routeName, null, ...params);
+  popTo = (routeName, data) => {
+    const params = filterParam(data);
+    this.dispatch({ type: ActionConst.POP_TO, routeName, params });
   };
 
-  popAndPush = (routeName, ...params) => {
-    this.run(ActionConst.POP_AND_PUSH, routeName, null, ...params);
+  popAndPush = (routeName, data) => {
+    const params = filterParam(data);
+    this.dispatch({ type: ActionConst.POP_AND_PUSH, routeName, params });
   };
 
-  replace = (routeName, ...params) => {
-    const res = uniteParams(routeName, params);
-    this.run(ActionConst.REPLACE, routeName, { key: routeName, index: 0, actions: [NavigationActions.navigate({
+  replace = (routeName, data) => {
+    const params = filterParam(data);
+    this.dispatch(NavigationActions.reset({ key: routeName, index: 0, actions: [NavigationActions.navigate({
       routeName,
-      params: res,
-    })] });
+      params,
+    })] }));
   };
 
-  reset = (routeName, ...params) => {
-    const res = uniteParams(routeName, params);
-    this.run(ActionConst.RESET, routeName, { key: null, index: 0, actions: [NavigationActions.navigate({
+  reset = (routeName, data) => {
+    const params = filterParam(data);
+    this.dispatch(NavigationActions.reset({ key: null, index: 0, actions: [NavigationActions.navigate({
       routeName,
-      params: res,
-    })] });
+      params,
+    })] }));
   };
 }
